v1.1.0 - Complete services + APK build
FIXES:
- All fees display working with static methods
- Type fixes for Dart 3.x compatibility
- Gradle 8.7 + AGP 8.3.0 + Kotlin 2.1.0

SERVICES:
- FeeService, StorageService, WalletService
- ChatService, JobService, SACCOService
- GameService, InternetSharingService

UI:
- All Fees screen in Profile tab

+++ pinc-network/pinc_network/lib/core/services/fee_service.dart
@@ -191,23 +191,23 @@ class FeeService {
   static double getFileStorageOverageFee() => 100.0;
   
   static Map<String, dynamic> calculateStorageFees(double usedTB) {
-    double freeLimit = getFreeStorageTB();
+    int freeLimit = getFreeStorageTB();
     if (usedTB <= freeLimit) {
       return {
         'usedTB': usedTB,
-        'freeLimit': freeLimit,
+        'freeLimit': freeLimit.toDouble(),
         'overageTB': 0.0,
         'fee': 0.0,
       };
     }
     
     double overageTB = usedTB - freeLimit;
-    double blocks = (overageTB / 10).ceil();
+    int blocks = (overageTB / 10).ceil();
     double fee = blocks * getFileStorageOverageFee();
     
     return {
       'usedTB': usedTB,
-      'freeLimit': freeLimit,
+      'freeLimit': freeLimit.toDouble(),
       'overageTB': overageTB,
       'fee': fee,
     };
